[x] 무한스크롤 구현
[x] 타입지정(any제거)

+++ server/src/module/follow/follow.service.ts
@@ -13,54 +13,58 @@ export class FollowService {
 		private userRepository: UserRepository
 	) {}
 
-	async postFollowing(user: User, followingNickname: string): Promise<Follow[]> {
+  //유저가 다른 닉네임을 팔로우 하기
+	async postFollowing(user: User, nickname: string): Promise<Follow[]> {
 		const findUser = await this.userRepository.find({
             where: {
-                nickname: followingNickname
+                nickname: nickname
             }
         })
-		if(user.id === findUser[0].id){
-			throw new BadRequestException(`Can not follow userself ${followingNickname}`)
+		if(user.nickname === findUser[0].nickname){
+			throw new BadRequestException(`Can not follow userself ${nickname}`)
 		}
-		return this.followRepository.postFollowing(user, followingNickname);
+		return this.followRepository.postFollowing(user, nickname);
 	}
 
-	async getFollowingList(nickname: string): Promise<Follow[]>{
+  //유저가 팔로우하고 있는 다른 유저들의 목록 리턴
+  async getFollowingList(nickname: string): Promise<Follow[]>{
 		//user_id가 user.id인 리스트
 		//user_id로 유저 테이블에서 유저 닉네임을 가져와야함.
-		const findUser = await this.userRepository.find({
-			where: {
-				nickname
-			}
-		})
+    const findUser = await this.userRepository.findOne({
+      where: {
+        nickname: nickname
+      }
+    });
 
-		const followingList = await this.followRepository.createQueryBuilder('follow')
+    //유저가 팔로우하고 있는 다른 유저들의 목록 조회
+    const followingList = await this.followRepository.createQueryBuilder('follow')
 			.select([
 				'follow.id AS id',
 				'follow.followingNickname AS followingNickname',
-				'user.nickname AS followerNickname'
+				'user.nickname AS userId'
 			])
 			.leftJoin(
 				'follow.user', 'user'
-			)
-			.where('follow.user_id = :user_id', {user_id: `${findUser[0].id}`})
-			.groupBy('follow.id')
-			.getRawMany()
-
+      )
+			.where('follow.user_id = :user_id', {user_id: `${findUser.id}`})
+      .getRawMany()
+    
 		return followingList;
-
-	}
-
-	async cancelFollowing(user: User, followingNickname: string): Promise<void> {
-		const result = await this.followRepository.delete({user, followingNickname});
+  }
+  
+  //유저가 특정 닉네임의 팔로우 중지
+  async cancelFollowing(user: User, nickname: string): Promise<void> {
+		const result = await this.followRepository.delete({user_id: user.id, followingNickname: nickname});
 
 		if(result.affected === 0) {
-			throw new NotFoundException(`Can't find Following with nickname ${followingNickname}`);
+			throw new NotFoundException(`Can't find Following with nickname ${nickname}`);
 		}
 	}
 
+  //유저를 팔로우하고 있는 다른 유저 목록 리턴
 	async getFollowerList(nickname: string): Promise<Follow[]> {
 		//user_id에 따른 nickname을 넣어야함
+    //유저를 팔로우하고 있는 다른 유저 목록 조회
 		const followerList = await this.followRepository.createQueryBuilder('follow')
 		.select([
 			'follow.id AS id',
@@ -72,23 +76,24 @@ export class FollowService {
 		)
 		.where('follow.followingNickname = :followingNickname', {followingNickname: `${nickname}`})
 		.groupBy('follow.id')
-		.getRawMany()
+      .getRawMany()
 
-		return followerList
+    return followerList;
 	}
 
-	async getFollowOrNot(user: User, userNickname: string): Promise<boolean> {
+  //유저가 특정 닉네임을 팔로우하고있는지 아닌지 판별
+  async getFollowOrNot(user: User, nickname: string): Promise<boolean> {
 		const FindExistOrNot = await this.followRepository.find({
 			where: {
 				user_id: user.id,
-				followingNickname: userNickname
+				followingNickname: nickname
 			}
 		})
 
 		if(FindExistOrNot.length !== 0) {
 			//팔로우 중
 			return true;
-		}
+    }
 		return false;
 	}
 }
